requirements.txt
openai>=1.0.0
pyyaml>=6.0
keyboard>=0.13.5
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.10.0
edge-tts>=6.1.0
faster-whisper>=0.10.0
mss>=9.0.0
Pillow>=10.0.0
pygame>=2.5.0
PyQt6>=6.5.0
